+++ src/main/java/io/github/mmalykhin/hmsproxy/routing/GetTableReqHandler.java
@@ -1,6 +1,7 @@
 package io.github.mmalykhin.hmsproxy.routing;
 
 import io.github.mmalykhin.hmsproxy.backend.CatalogBackend;
+import io.github.mmalykhin.hmsproxy.thriftbridge.ThriftValueConverter;
 import java.lang.reflect.Method;
 import org.apache.hadoop.hive.metastore.api.GetTableRequest;
 
@@ -13,17 +14,31 @@ final class GetTableReqHandler implements SpecialCaseHandler {
 
   @Override
   public Object handle(Method method, Object[] args) throws Throwable {
-    GetTableRequest request = (GetTableRequest) args[0];
-    CatalogRouter.ResolvedNamespace namespace = support.federationLayer.resolveRequestNamespace(
-        request.getCatName(), request.getDbName());
+    Object request = args[0];
+    String catName = ThriftReflectionCache.readString(request, "catName", "getCatName");
+    String dbName = ThriftReflectionCache.readString(request, "dbName", "getDbName");
+    String tblName = ThriftReflectionCache.readString(request, "tblName", "getTblName");
+
+    CatalogRouter.ResolvedNamespace namespace = support.federationLayer.resolveRequestNamespace(catName, dbName);
     RequestContext.currentObservation().recordNamespace(namespace);
-    support.recordDefaultCatalogRouteIfImplicit(method.getName(), request.getCatName(), request.getDbName(), namespace);
+    support.recordDefaultCatalogRouteIfImplicit(method.getName(), catName, dbName, namespace);
     CatalogBackend backend = namespace.backend();
     support.validateExposedDatabaseAccess(method.getName(), namespace);
-    support.validateExposedTableAccess(method.getName(), namespace, request.getTblName());
-    GetTableRequest routedRequest =
-        (GetTableRequest) support.federationLayer.internalizeTableRequest(request, namespace);
-    Object result = support.invokeViaRequest(backend, routedRequest, method.getName());
+    support.validateExposedTableAccess(method.getName(), namespace, tblName);
+
+    Object result;
+    if (backend.runtimeProfile().isHive4() && !(request instanceof GetTableRequest)) {
+      Object routedRequest = support.federationLayer.internalizeArgument(request, namespace);
+      result = support.invokeBackendNamed(backend, "get_table_req", routedRequest);
+    } else {
+      GetTableRequest apacheReq = request instanceof GetTableRequest standardReq
+          ? standardReq
+          : (GetTableRequest) ThriftValueConverter.convertTBase(request, GetTableRequest.class);
+      GetTableRequest routedRequest =
+          (GetTableRequest) support.federationLayer.internalizeTableRequest(apacheReq, namespace);
+      result = support.invokeViaRequest(backend, routedRequest, method.getName());
+    }
+
     result = support.filterSingleTableResult(method.getName(), namespace, result);
     return support.federationLayer.externalizeResult(result, namespace);
   }